Skip evaluating shape keys where possible
If a shape key is the same as the reference key, then the evaluated
reference key can be used for that shape key also.

If a shape key only affects vertices which are unaffected by the new
pose, then the parts unaffected by the shape key can be set to the
evaluated reference key.